DPDK Driver 2022.1.3 patch: Added support for PF/VF 4k queues

+++ QDMA/DPDK/drivers/net/qdma/qdma.h
@@ -1,7 +1,8 @@
 /*-
  * BSD LICENSE
  *
- * Copyright(c) 2017-2022 Xilinx, Inc. All rights reserved.
+ * Copyright (c) 2017-2022 Xilinx, Inc. All rights reserved.
+ * Copyright (c) 2022, Advanced Micro Devices, Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -199,7 +200,7 @@ struct qdma_rx_queue {
 
 	enum rte_pmd_qdma_bypass_desc_len	bypass_desc_sz:7;
 	uint8_t			func_id; /**< RX queue index. */
-	uint32_t		ep_addr;
+	uint64_t		ep_addr;
 
 	int8_t			ringszidx;
 	int8_t			cmpt_ringszidx;
@@ -252,7 +253,7 @@ struct qdma_tx_queue {
 
 	struct qdma_pkt_stats stats;
 
-	uint32_t			ep_addr;
+	uint64_t			ep_addr;
 	uint32_t			queue_id; /* TX queue index. */
 	uint32_t			num_queues; /* TX queue index. */
 	const struct rte_memzone	*tx_mz;
