+++ QDMA/DPDK/drivers/net/qdma/qdma_rxtx.h
@@ -1,7 +1,8 @@
 /*-
  * BSD LICENSE
  *
- * Copyright(c) 2019-2022 Xilinx, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Xilinx, Inc. All rights reserved.
+ * Copyright (c) 2022, Advanced Micro Devices, Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -41,8 +42,8 @@ void qdma_get_device_info(void *queue_hndl,
 		enum qdma_ip_type *ip_type);
 struct qdma_ul_st_h2c_desc *get_st_h2c_desc(void *queue_hndl);
 struct qdma_ul_mm_desc *get_mm_h2c_desc(void *queue_hndl);
-uint32_t get_mm_c2h_ep_addr(void *queue_hndl);
-uint32_t get_mm_h2c_ep_addr(void *queue_hndl);
+uint64_t get_mm_c2h_ep_addr(void *queue_hndl);
+uint64_t get_mm_h2c_ep_addr(void *queue_hndl);
 uint32_t get_mm_buff_size(void *queue_hndl);
 
 #endif /* QDMA_DPDK_RXTX_H_ */
